Move sync scheduling to the login activity
