+++ app/src/main/java/org/fundacionparaguaya/advisorapp/data/JobCreator.java
@@ -5,6 +5,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import com.evernote.android.job.Job;
+import com.evernote.android.job.JobRequest;
 
 import org.fundacionparaguaya.advisorapp.AdvisorApplication;
 import org.fundacionparaguaya.advisorapp.repositories.SyncManager;
@@ -35,6 +36,26 @@ public class JobCreator implements com.evernote.android.job.JobCreator {
                 return new SyncJob(mSyncManager);
             default:
                 return null;
+        }
+    }
+
+    public class SyncJob extends Job {
+        public static final String TAG = "SyncJob";
+
+        private SyncManager mSyncManager;
+
+        public SyncJob(SyncManager syncManager) {
+            super();
+            this.mSyncManager = syncManager;
+        }
+
+        @Override
+        @NonNull
+        protected Result onRunJob(@NonNull Params params) {
+            if (mSyncManager.sync())
+                return Result.SUCCESS;
+            else
+                return Result.FAILURE;
         }
     }
 }
